@@ -35,7 +35,6 @@
   },
   "dependencies": {
     "file-url": "^2.0.0",
-    "iconv-lite": "^0.4.15",
     "minimatch": "^3.0.3",
     "moment": "^2.17.1",
     "url-relative": "^1.0.0",
@@ -45,26 +44,27 @@
     "xmldom": "^0.1.22"
   },
   "devDependencies": {
+    "@sourcegraph/eslint-config": "^0.20.19",
     "@types/chai": "4.0.6",
     "@types/chai-as-promised": "^7.1.0",
     "@types/file-url": "^2.0.0",
     "@types/minimatch": "^3.0.1",
     "@types/mocha": "^2.2.44",
-    "@types/node": "7.0.32",
+    "@types/node": "^14.14.14",
     "@types/semver": "^5.4.0",
     "@types/xmldom": "^0.1.29",
     "chai": "^4.0.2",
     "chai-as-promised": "^7.0.0",
+    "eslint": "^7.16.0",
     "husky": "^0.14.3",
+    "iconv-lite": "^0.6.2",
     "mocha": "^5.0.0",
     "nyc": "^15.1.0",
-    "prettier": "1.14.2",
+    "prettier": "^2.2.1",
     "semantic-release": "^17.3.0",
     "semantic-release-vsce": "^3.0.1",
     "semver": "^5.3.0",
-    "tslint": "^5.4.3",
-    "tslint-config-prettier": "^1.6.0",
-    "typescript": "^2.1.4",
+    "typescript": "^4.1.3",
     "validate-commit-msg": "^2.14.0",
     "vsce": "^1.28.0",
     "vscode-debugadapter-testsupport": "^1.11.0"
@@ -95,8 +95,9 @@
     "start": "code testproject --extensionDevelopmentPath=.",
     "test": "mocha out/test --timeout 20000 --slow 1000 --retries 4",
     "cover": "nyc --all mocha out/test --timeout 20000 --slow 1000 --retries 4",
-    "lint": "npm run tslint && npm run prettier",
-    "tslint": "tslint -c tslint.json 'src/**/*.ts'",
+    "lint": "npm run eslint && npm run prettier",
+    "eslint": "eslint \"src/**/*.ts\"",
+    "eslint-watch": "npm run eslint -- --watch",
     "prettier": "prettier \"**/{*.json,*.yml,.*.yml,*.ts,.prettierrc,*.md}\" --write --list-different",
     "semantic-release": "semantic-release",
     "commitmsg": "validate-commit-msg"
