@@ -1,4 +1,5 @@
-/*---------------------------------------------------------------------------------------------
+/* eslint no-sync: warn */
+/* ---------------------------------------------------------------------------------------------
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
@@ -7,15 +8,16 @@ import * as Path from 'path'
 import * as FS from 'fs'
 import * as CP from 'child_process'
 
-export class Terminal {
+// eslint-disable-next-line @typescript-eslint/no-extraneous-class
+export abstract class Terminal {
     private static _terminalService: ITerminalService
 
     public static launchInTerminal(
-        dir: string,
+        directory: string,
         args: string[],
-        envVars: { [key: string]: string }
-    ): Promise<CP.ChildProcess> {
-        return this.terminalService().launchInTerminal(dir, args, envVars)
+        environmentVars: { [key: string]: string | undefined }
+    ): Promise<CP.ChildProcess | undefined> {
+        return this.terminalService().launchInTerminal(directory, args, environmentVars)
     }
 
     public static killTree(processId: number): Promise<any> {
@@ -46,7 +48,11 @@ export class Terminal {
 }
 
 interface ITerminalService {
-    launchInTerminal(dir: string, args: string[], envVars: { [key: string]: string }): Promise<CP.ChildProcess>
+    launchInTerminal(
+        directory: string,
+        args: string[],
+        environmentVars: { [key: string]: string | undefined }
+    ): Promise<CP.ChildProcess | undefined>
     killTree(pid: number): Promise<any>
     isOnPath(program: string): boolean
 }
@@ -54,24 +60,28 @@ interface ITerminalService {
 class DefaultTerminalService implements ITerminalService {
     protected static TERMINAL_TITLE = 'VS Code Console'
 
-    public launchInTerminal(dir: string, args: string[], envVars: { [key: string]: string }): Promise<CP.ChildProcess> {
+    public launchInTerminal(
+        directory: string,
+        args: string[],
+        environmentVars: { [key: string]: string }
+    ): Promise<CP.ChildProcess | undefined> {
         throw new Error('launchInTerminal not implemented')
     }
 
     public killTree(pid: number): Promise<any> {
         // on linux and OS X we kill all direct and indirect child processes as well
 
-        return new Promise<any>((resolve, reject) => {
+        return new Promise<void>((resolve, reject) => {
             try {
-                const cmd = Path.join(__dirname, './terminateProcess.sh')
-                const result = (<any>CP).spawnSync(cmd, [pid.toString()])
+                const command = Path.join(__dirname, './terminateProcess.sh')
+                const result = CP.spawnSync(command, [pid.toString()])
                 if (result.error) {
                     reject(result.error)
                 } else {
                     resolve()
                 }
-            } catch (err) {
-                reject(err)
+            } catch (error) {
+                reject(error)
             }
         })
     }
@@ -100,26 +110,30 @@ class DefaultTerminalService implements ITerminalService {
 class WindowsTerminalService extends DefaultTerminalService {
     private static CMD = 'cmd.exe'
 
-    public launchInTerminal(dir: string, args: string[], envVars: { [key: string]: string }): Promise<CP.ChildProcess> {
+    public launchInTerminal(
+        directory: string,
+        args: string[],
+        environmentVars: { [key: string]: string }
+    ): Promise<CP.ChildProcess> {
         return new Promise<CP.ChildProcess>((resolve, reject) => {
-            const title = `"${dir} - ${WindowsTerminalService.TERMINAL_TITLE}"`
+            const title = `"${directory} - ${WindowsTerminalService.TERMINAL_TITLE}"`
             const command = `""${args.join('" "')}" & pause"` // use '|' to only pause on non-zero exit code
 
-            const cmdArgs = ['/c', 'start', title, '/wait', 'cmd.exe', '/c', command]
+            const commandArguments = ['/c', 'start', title, '/wait', 'cmd.exe', '/c', command]
 
             // merge environment variables into a copy of the process.env
-            const env = extendObject(extendObject({}, process.env), envVars)
+            const environment = extendObject(extendObject({}, process.env), environmentVars)
 
             const options: any = {
-                cwd: dir,
-                env: env,
+                cwd: directory,
+                env: environment,
                 windowsVerbatimArguments: true,
             }
 
-            const cmd = CP.spawn(WindowsTerminalService.CMD, cmdArgs, options)
-            cmd.on('error', reject)
+            const command_ = CP.spawn(WindowsTerminalService.CMD, commandArguments, options)
+            command_.on('error', reject)
 
-            resolve(cmd)
+            resolve(command_)
         })
     }
 
@@ -127,13 +141,13 @@ class WindowsTerminalService extends DefaultTerminalService {
         // when killing a process in Windows its child processes are *not* killed but become root processes.
         // Therefore we use TASKKILL.EXE
 
-        return new Promise<any>((resolve, reject) => {
-            const cmd = `taskkill /F /T /PID ${pid}`
+        return new Promise<void>((resolve, reject) => {
+            const command = `taskkill /F /T /PID ${pid}`
             try {
-                CP.execSync(cmd)
+                CP.execSync(command)
                 resolve()
-            } catch (err) {
-                reject(err)
+            } catch (error) {
+                reject(error)
             }
         })
     }
@@ -143,47 +157,49 @@ class LinuxTerminalService extends DefaultTerminalService {
     private static LINUX_TERM = '/usr/bin/gnome-terminal' // private const string LINUX_TERM = "/usr/bin/x-terminal-emulator";
     private static WAIT_MESSAGE = 'Press any key to continue...'
 
-    public launchInTerminal(dir: string, args: string[], envVars: { [key: string]: string }): Promise<CP.ChildProcess> {
-        return new Promise<CP.ChildProcess>((resolve, reject) => {
+    public launchInTerminal(
+        directory: string,
+        args: string[],
+        environmentVars: { [key: string]: string }
+    ): Promise<CP.ChildProcess | undefined> {
+        return new Promise<CP.ChildProcess | undefined>((resolve, reject) => {
             if (!FS.existsSync(LinuxTerminalService.LINUX_TERM)) {
                 reject(
                     new Error(
-                        `Cannot find '${
-                            LinuxTerminalService.LINUX_TERM
-                        }' for launching the node program. See http://go.microsoft.com/fwlink/?linkID=534832#_20002`
+                        `Cannot find '${LinuxTerminalService.LINUX_TERM}' for launching the node program. See http://go.microsoft.com/fwlink/?linkID=534832#_20002`
                     )
                 )
                 return
             }
 
-            const bashCommand = `cd "${dir}"; "${args.join('" "')}"; echo; read -p "${
+            const bashCommand = `cd "${directory}"; "${args.join('" "')}"; echo; read -p "${
                 LinuxTerminalService.WAIT_MESSAGE
             }" -n1;`
 
-            const termArgs = [
+            const termArguments = [
                 '--title',
                 `"${LinuxTerminalService.TERMINAL_TITLE}"`,
                 '-x',
                 'bash',
                 '-c',
-                `\'\'${bashCommand}\'\'`, // wrapping argument in two sets of ' because node is so "friendly" that it removes one set...
+                `''${bashCommand}''`, // wrapping argument in two sets of ' because node is so "friendly" that it removes one set...
             ]
 
             // merge environment variables into a copy of the process.env
-            const env = extendObject(extendObject({}, process.env), envVars)
+            const environment = extendObject(extendObject({}, process.env), environmentVars)
 
             const options: any = {
-                env: env,
+                env: environment,
             }
 
-            const cmd = CP.spawn(LinuxTerminalService.LINUX_TERM, termArgs, options)
-            cmd.on('error', reject)
-            cmd.on('exit', (code: number) => {
+            const command = CP.spawn(LinuxTerminalService.LINUX_TERM, termArguments, options)
+            command.on('error', reject)
+            command.on('exit', (code: number) => {
                 if (code === 0) {
                     // OK
-                    resolve() // since cmd is not the terminal process but just a launcher, we do not pass it in the resolve to the caller
+                    resolve(undefined) // since cmd is not the terminal process but just a launcher, we do not pass it in the resolve to the caller
                 } else {
-                    reject(new Error('exit code: ' + code))
+                    reject(new Error(`exit code: ${code}`))
                 }
             })
         })
@@ -193,36 +209,40 @@ class LinuxTerminalService extends DefaultTerminalService {
 class MacTerminalService extends DefaultTerminalService {
     private static OSASCRIPT = '/usr/bin/osascript' // osascript is the AppleScript interpreter on OS X
 
-    public launchInTerminal(dir: string, args: string[], envVars: { [key: string]: string }): Promise<CP.ChildProcess> {
-        return new Promise<CP.ChildProcess>((resolve, reject) => {
+    public launchInTerminal(
+        directory: string,
+        args: string[],
+        environmentVars: { [key: string]: string }
+    ): Promise<CP.ChildProcess | undefined> {
+        return new Promise<CP.ChildProcess | undefined>((resolve, reject) => {
             // first fix the PATH so that 'runtimePath' can be found if installed with 'brew'
             // Utilities.FixPathOnOSX();
 
             // On OS X we do not launch the program directly but we launch an AppleScript that creates (or reuses) a Terminal window
             // and then launches the program inside that window.
 
-            const osaArgs = [
+            const osaArguments = [
                 Path.join(__dirname, './terminalHelper.scpt'),
                 '-t',
                 MacTerminalService.TERMINAL_TITLE,
                 '-w',
-                dir,
+                directory,
             ]
 
             for (const a of args) {
-                osaArgs.push('-pa')
-                osaArgs.push(a)
+                osaArguments.push('-pa')
+                osaArguments.push(a)
             }
 
-            if (envVars) {
-                for (const key in envVars) {
-                    osaArgs.push('-e')
-                    osaArgs.push(key + '=' + envVars[key])
+            if (environmentVars) {
+                for (const key of Object.keys(environmentVars)) {
+                    osaArguments.push('-e')
+                    osaArguments.push(key + '=' + environmentVars[key])
                 }
             }
 
             let stderr = ''
-            const osa = CP.spawn(MacTerminalService.OSASCRIPT, osaArgs)
+            const osa = CP.spawn(MacTerminalService.OSASCRIPT, osaArguments)
             osa.on('error', reject)
             osa.stderr.on('data', (data: Buffer) => {
                 stderr += data.toString()
@@ -230,13 +250,11 @@ class MacTerminalService extends DefaultTerminalService {
             osa.on('exit', (code: number) => {
                 if (code === 0) {
                     // OK
-                    resolve() // since cmd is not the terminal process but just the osa tool, we do not pass it in the resolve to the caller
+                    resolve(undefined) // since cmd is not the terminal process but just the osa tool, we do not pass it in the resolve to the caller
+                } else if (stderr) {
+                    reject(new Error(stderr))
                 } else {
-                    if (stderr) {
-                        reject(new Error(stderr))
-                    } else {
-                        reject(new Error('exit code: ' + code))
-                    }
+                    reject(new Error(`exit code: ${code}`))
                 }
             })
         })
@@ -246,9 +264,9 @@ class MacTerminalService extends DefaultTerminalService {
 // ---- private utilities ----
 
 function extendObject<T>(objectCopy: T, object: T): T {
-    for (let key in object) {
-        if (object.hasOwnProperty(key)) {
-            ;(<any>objectCopy)[key] = (<any>object)[key]
+    for (const key in object) {
+        if (Object.prototype.hasOwnProperty.call(object, key)) {
+            ;(objectCopy as any)[key] = (object as any)[key]
         }
     }
 
